# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FuelGrid OS API
  version: 0.9.0
  description: |
    Hand-maintained spec for the FuelGrid OS API surface. Kept in
    lockstep with the Go handlers by review until Stage 11+ wires
    automatic generation from chi route metadata.

    The Stage-10 CI workflow validates this file with `redocly lint`.
    SDK regeneration is manual today — see packages/sdk/.
servers:
  - url: http://localhost:8080
    description: Local dev (`go run ./services/api/cmd/api`)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/branding/logo:
    get:
      tags: [Branding]
      summary: Stream the tenant logo
      description: Returns the stored logo with its image content type, or 404 when none is set.
      responses:
        '200':
          description: OK
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Branding]
      summary: Upload the tenant logo
      description: >-
        Multipart upload of a PNG or JPEG image, capped at 1 MiB. Requires the
        companies.manage permission. Non-image content is rejected 400; an
        oversized body is rejected 413.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Branding' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { description: Image exceeds the 1 MiB size cap }
    delete:
      tags: [Branding]
      summary: Clear the tenant logo
      description: Requires the companies.manage permission. Idempotent (204 even when no logo was set).
      responses:
        '204': { description: Cleared }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    BadRequest:
      description: Validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: Authenticated but not authorized
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error:  { type: string }
        status: { type: integer }
    # Generic JSON object for endpoints whose payload is documented for
    # coverage but whose field-level shape is intentionally left open.
...
    Branding:
      type: object
      properties:
        display_name:      { type: string }
        legal_name:        { type: string }
        tax_id:            { type: string }
        registration_no:   { type: string }
        address_line1:     { type: string }
        address_line2:     { type: string }
        city:              { type: string }
        country:           { type: string }
        phone:             { type: string }
        email:             { type: string }
        website:           { type: string }
        footer_note:       { type: string }
        has_logo:          { type: boolean }
        logo_content_type: { type: string }
        logo_url:          { type: string }
        updated_at:        { type: string, format: date-time }
...